update upload file mode

+++ application/controllers/bakapi.php
@@ -158,6 +158,9 @@ class Imgapi extends CI_Controller {
                 $imgurl = 'cache/images/ed2k'.basename($imgurl);
                 file_put_contents($imgurl, $html);
                 chmod($imgurl, 0777);
+                if( !file_exists($imgurl) || filesize($imgurl) < 2000){
+                  die('0');
+                }
                 $imginfo['md5'] = md5_file($imgurl);
                 unlink($imgurl);
 //exit;//var_dump($imginfo);exit;
